// src/types/about.ts
// define about page content interface
export interface AboutPageContentProps {
  introMainTitle: string;
  introFirstSectionDescription: string;
  introSecSectionDescription: string;
  introThirdSectionDescription: string;
  introBoldFirstSectionDescription: string;
  introBoldSecondSectionDescription: string;
  introFirstSideImgUrl: string;
  introSecSideImgUrl: string;
  secIntroTitle: string;
}

// define our values interface
export interface OurValuesProps {
  title: string;
  imgUrl: string;
  description: string;
}

// define our mission and vision interface
export interface OurMissionVisionProps {
  title: string;
  secTitle: string;
  imgUrl: string;
  description: string;
}